org-protocol default template should be nil
Hi Carsten,

this little patch fixes an issue Richard brought up.

We always used the "w" template as the default for `org-remember' and
also used it for `org-capture' for historical reasons.

Unfortunately, this breaks, if the user has no "w" template defined.

The patch below simply set's the custom variable
`org-protocol-default-template-key' to nil, so the interactive template
selection is used by default.  This works for both, remember an capture.

I will adjust the docs, once the patch is applied.
Thanks,

  Sebastian