Compiling multiple times the LaTeX output
* lisp/org-latex.el (org-latex-to-pdf-process): Add a third pdflatex run.

Hi Carsten,

Carsten Dominik wrote:
> I have just pushed a fix which will use texi2dvi when available. The echo
> area will now also give an indication of the kinds of errors that happened
> during the final pdflatex run.

Thanks to Matthew for the info about texi2dvi!

> This is a big improvement, thanks to Sebastian for kicking off this
> discussion, and to all who contributed.

You're welcome. It was needed for me as well...

> Please test the new processing and let me know if there are problems.

Tested on the document for which I experienced the troubles in the first place
(because my TOC was more than one page long). Works perfect!

Only one comment, after reading your code in org-latex.el: if there is no
texi2dvi executable, please run pdflatex 3 times, as this is the right number
to get the TOC in sync' with the document, when it's more than one page long
(at least, in my above case, with a couple of org-babel sql source blocks).

--8<---------------cut here---------------start------------->8---
Changes in HEAD
	Modified lisp/org-latex.el

--8<---------------cut here---------------end--------------->8---

Thanks.

Best regards,
  Seb