correct doc typos
Hi

Here is a patch for some proofreading corrections for the Org manual.
This is just an sample to check if it is in suitable format.  We have
a lot more corrections to come.

Because there are more than 20 lines affected I'm assuming we will
need to do copyright assignments.  If not I can send all the patches
now.  Alternatively would you prefer us to wait until the assignment
is done?

I haven't included a changelog entry since these are just for typos
etc and don't make any major changes to the meaning of the text.

Credit to my colleague Barry Gidden for the proofreading, he did the
real work on this.