How to write special LaTeX symbols
Xin Shi <shixin111@gmail.com> wrote:

>
> Shr\"odinger   will give the o with two dots on top. Notice that the " is a double quotation mark. 
>
> However, when org translate that into LaTeX, it will become two single quotation mark! \'' (it's very hard to see the difference, but the pdf version will
> see the wrong result).
>
> Could someone tell me how to do that?
>

It works correctly in headlines, but not in running text. I think [fn:1]
that it is a bug and that the following patch fixes it:

--8<---------------cut here---------------start------------->8---
--8<---------------cut here---------------end--------------->8---

but I have not tested it extensively and it may do more harm than good:
I'd wait for a more definitive opinion.

Alternatively, you can use UTF-8 in your org file and write Schrödinger
explicitly.  This will survive the LaTeX export intact and the
\usepackage[utf8]{inputenc} in the LaTeX file will do the right thing
with it.[fn:2]

Nick

Footnotes:

[fn:1] but I'm really not sure: I've lost track of how things work in
LaTeX export - sigh...

[fn:2] I'm not sure whether it will survive the email trip
though. Here's hoping that it will.