PATCH: Support `org-export-process-option-filters' in latex export
`org-export-process-option-filters' is supported for HTML export but not
for latex export. The patch at the bottom fixes that.

I note that there is one other location in the same function where it
seems the same change could be made. I suspect this is not necessary,
but if someone is familiar with the org-latex code they might like to
have a look. (I didn't manage to understand the roles of the several
variables with names like *-opt-plist *-options-plist in that function.)

I'm protecting this patch from patchwork with ^$; it's the patch at the
bottom that I propose.

$ diff --git a/lisp/org-latex.el b/lisp/org-latex.el
$ index 51ee6d2..53d6b40 100644
$ --- a/lisp/org-latex.el
$ +++ b/lisp/org-latex.el
$ @@ -700,7 +700,8 @@ when PUB-DIR is set, use this as the publishing directory."
$  			     '(:org-license-to-kill nil))))
$    (org-update-radio-target-regexp)
$    (org-export-latex-set-initial-vars ext-plist arg)
$ -  (setq org-export-opt-plist org-export-latex-options-plist)
$ +  (setq org-export-opt-plist
$ +  	(org-export-process-option-filters org-export-latex-options-plist))
$    (org-install-letbind)
$    (run-hooks 'org-export-latex-after-initial-vars-hook)
$    (let* ((wcf (current-window-configuration))
$

    Support `org-export-process-option-filters' in latex export

    * lisp/org-latex.el (org-export-as-latex): Process export property
    list with `org-export-process-option-filters' early in latex export

	Modified lisp/org-latex.el