Call kill-buffer with argument nil
* ob-haskell.el (org-babel-haskell-export-to-lhs): Call kill-buffer
with argument indiciating to kill current buffer. Emacs 22
compatibility.

GNU Emacs 22.3.2 (i686-pc-linux-gnu) of 2011-05-28 on x60s

C-h f kill-buffer RET

kill-buffer is an interactive built-in function in `C source code'.
It is bound to C-x k.
(kill-buffer buffer)

Kill the buffer buffer.
The argument may be a buffer or the name of a buffer.
With a nil argument, kill the current buffer.